Extract code related to --version flag to flag_version.go
